requirements.txt
blinker==1.7.0
boto3==1.28.21
botocore==1.31.21
click==8.1.7
colorama==0.4.6
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
pipreq==0.4
protobuf==4.21.12
psycopg2==2.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
Werkzeug==3.0.1
jq==1.6.0
langchain==0.0.350
langchain-community==0.0.2
langchain-core==0.1.0
langsmith==0.0.69
meilisearch==0.28.4
oauthlib==3.2.0
openai==1.3.9
regex==2023.10.3
requests==2.31.0
tiktoken==0.5.2